refactor: deprecated document_index.py (moved method to "group.py")
